Update SAI ozone safe start kpetlock4.03.2023.py
I just put the new hypothetical k values in, which are proportional to the corresponding hypothetical  SAD values, as the SAD is just a direct multiplier (1 to 1) in the Wegner formula for k.  Similarly, looks like the changes in dO3 value will be proportional (1 to 1) to the hypothetical changes (error bar scenarios) in k value, as k is just a direct multiplier in the emulator equation for dO3.  I'm not sure how we want to format these hypotheticals in the code, so I just left it as alternate k values for now, (based on the alternate SAD's), which can be substituted into the equation for dO3, and the addition of these error scenarios to the plot I guess is a work in progress.. 
Kyle