Update SAI ozone safe start kpetlock4.03.2023.py
Hi Dan, 
I converted cgas to cm/s (as cgsc) to be compatible with the air volume units (cm^3) in the SAD units (one side of the cube) in the calculation of K.  (I don't know if that follows convention but it seems to make sense in the unit analysis for the calculation), which brings the O3 depletion results back to the order of 10^5.  
Also, I noticed several parameters ported (prefaced) with prm which don't seem to be defined in the code (i.e. clp, dcl, dae and 'year'), -just wondering how that works.  Also, I activated a for loop for the sadl data in my code on my end (in my editor) which is working.  I can put this into the code on github, unless we have it working with param.  Kyle