Update SAI ozone safe start kpetlock4.03.2023.py
Hi Dan, 
I see now why you multiplied dae by da20, as dae from the Arise paper is in ug/kg, as opposed to a mass/volume unit, as the Richter SO4 data is.  
That said, I'm more comfortable using the Richter SO4 data, as I have concerns about the scenario of continuously increasing aerosol injection rates to compensate for increasing GHG emissions.  
I activated the ksal 'for' loop to produce ksl, a list of reaction rates based on the CESM SAD projection in Tilmes et al., but I didn't substitute it into the emulator equation yet for O3 depletion, as it seems it would get messy as you predicted, with multiple variables.  
I saw the parameters you put into the param file so far.  I thought of reading the successive element lines from the param data matrix of parameter value lists, using 'readlines' and then 'line.split', but I haven't done that in a while and I'm not sure if that would work for this, so it may be better to use param the way you had in mind, which I'm still trying to learn more about.  
Also, I was trying to find that source you referred me to with the Cl dissipation data curve (in green) as it looked like the Cl dissipation rate shown there from 2023 going forward is significantly greater than the -8.2 PPT/yr that I had in the code from the earlier source.  Looked like the newer data showed 
more like -50 PPT/yr.  I can adjust it, I'm just having trouble finding that data source now.  
Kyle