Update SAI ozone safe start kpetlock4.03.2023.py
Hi Dan, 
Thought I made this change before, but I put mean molecular speed in terms of cm/s to be compatible with the other dimensional units in the calculation of K.  This (theoretically) would put the reaction rate (k) in terms of rxns/cm^3 sec, which of course would increase the 03 depletion results by 2 orders of magnitude.  Please let me know if this isn't right.  
Kyle