Update SAI ozone safe start kpetlock4.03.2023.py
Been wanting to discuss this with you and get your perspective:
Reviewing the emulator equation, it seemed to me like the reactants should be in terms of molecules/unit volume, to be compatible with the reaction rate, in rxns/unit vol - sec, to produce molec/unit vol - s of product (including Cl2 => 2Cl), to produce molec/unit vol -s of O3 destroyed through the R1 pathway.   This is why I put in gms and mds, to calculate the molecular density of added SO4, for now anyway.